views: make the live leg answer with what it observed, and with the right instant
Review of the first pass found defects in every layer of the live leg: what the
generated file CLAIMS, what it RETURNS, and whether it binds at all. Fixes in
that order.

What it returns
---------------
The index leg emitted event_timestamp bare. DuckDB reads the archives' Parquet
column as TIMESTAMP WITH TIME ZONE and the index's DATETIME as a naive
TIMESTAMP, and the union reconciles that pair by reading the naive value in the
READER's session timezone. The index stores UTC, so on any non-UTC session
every index row landed off by that offset: the same event answering with two
different instants depending on which leg served it, and a UTC range filter
returning nothing for an event inside it. That is the silent gap this flag
exists to close, reintroduced on the fresh side. It now reads
"event_timestamp" AT TIME ZONE 'UTC'. The strftimes that derive the Hive
columns stay on the naive value, where they are already correct.

The overlap went to the index. Both legs carry the same event, but not the same
knowledge: an archived row knows its bintrail_id, event_date and event_hour
from its path, and an index row derives or forgoes them. Handing it to the
index replaced a known source with NULL for every event in the overlap window,
so a WHERE bintrail_id = ... missed rows the archives hold, and a GROUP BY
bintrail_id was wrong for that whole window. The archives now win it. This is
where the SQL deliberately diverges from query.MergeResults: there the two
copies carry identical fields and the winner is immaterial, here it is not.

event_date became VARCHAR when the flag was on, because strftime returns
VARCHAR and the archives' hive column is DATE. Equality survived the implicit
cast; date_trunc, - INTERVAL and comparisons against current_date did not. Cast
to DATE. event_hour needs no cast — the hive column reads back VARCHAR too.

event_type_code is cast to INTEGER, not the index's TINYINT: INTEGER is what
the archives read back and what the archives-only file has always produced, and
a narrowing cast would make an unknown future event type fail the whole query,
which is the opposite of what the decoded CASE's ELSE branch promises.

Whether it binds
----------------
The index leg named every column of archive.BinlogEventColumns with nothing
checking they exist. Against an index migrated to an earlier point than this
build's schema — the console sets EnsureSchema: false and never migrates
registry servers — DuckDB answers with a binder error naming an internal column
and creates NO events view at all. The command now reads the live table's
column set and emits NULL for what is absent, which is the cold leg's
union_by_name applied to the leg that has no such mechanism. Not best effort,
unlike attribution: an unreadable column list or a missing binlog_events fails
the command with a message instead of producing a file that cannot bind.

With no archive source registered the events view was skipped entirely, so
--include-live on an index that has never rotated emitted a preamble, asked for
a password, attached read-only and then defined nothing. That is the state a
fresh install is in, and the one where the live leg is worth most. It now emits
the index leg alone. closes #1485

What it claims
--------------
Attribution was one COUNT(*) with err == nil && n == 1, so a file-mode index
(which registers no server and serves exactly ONE source), an index too old to
have bintrail_servers, an account without SELECT on it and a dropped connection
all produced the same sentence: "index serves more than one source; not
attributable per row". False for all but the first, and 1146 got the opposite
verdict from the one status.knownSourceCount gives the identical input. The
attribution is now four states that each say what was OBSERVED and never why
(that a zero count means a file-mode index is an inference, and it stays out of
the artifact). None of the error branches had a test; all of them do now.

The probed id was written into the file with nothing checking it against the
bintrail_id= path segments the cold leg reads, which rotate takes verbatim from
--bintrail-id and never validates. On a disagreement one source appears as two
servers in one view. Cross-checked now, and a disagreement falls back to
unattributed naming both values. An explicit --bintrail-id wins and skips the
probe: it already scoped the archive paths in this same file.

The archives-only note said "Regenerate with a reachable index", naming no
flag, so an operator who already passed --index-dsn regenerated and got a
byte-identical file. It names --include-live now. The console download never
sets LiveIndex and has no such flag, so it carries wording that is true there
instead (Input.LiveLegUnavailable).

A unix-socket --index-dsn was refused by config.ParseSourceDSN, whose message
names --source-dsn and justifies itself with binlog replication: the wrong flag
and an irrelevant reason for a command that generates text. It parses its own
DSN now. Still refused, deliberately: DuckDB's mysql extension does accept a
SOCKET, but this file locates the index by host and port precisely so it can be
run from another machine, and a socket path names only one. A loopback host is
emitted with a comment rather than refused — it is usable where it was
generated, and a reader elsewhere needs to know it is not a location.

Generate and GenerateViews had different preconditions on one Input:
GenerateViews emits no preamble, so a LiveIndex reaching it rendered a view
over a catalog nothing attaches. It drops the leg, which makes the divergence
impossible rather than documented.

Added: the cost an operator would otherwise measure and blame on the view. A
filter does not reach the index — the derived Hive columns are computed above
the scan and the anti-join needs the archives regardless — so every query
streams the live binlog_events whole. The note points at the attached table for
narrow reads.

Tests: an executable DuckDB test runs the two-leg view in a session set to
America/Bogota against a stand-in catalog typed the way the mysql extension
delivers the index's own DDL. On a UTC box the timestamp bug is invisible,
which is how it shipped. Seventeen mutations verified red first, one per fix.

+++ internal/views/testdata/views.golden.sql
@@ -42,8 +42,9 @@ CREATE OR REPLACE SECRET bintrail_s3_chain (TYPE s3, PROVIDER credential_chain, 
 --
 -- SCOPE: these are the ARCHIVED events only. Partitions rotation has not
 -- archived yet exist solely in the index, so the most recent window is
--- absent here and reads as if nothing happened. Regenerate with a
--- reachable index to add the live leg.
+-- absent here and reads as if nothing happened.
+-- Add a leg over the index by regenerating with --include-live:
+--   bintrail views --index-dsn ... --include-live
 CREATE OR REPLACE VIEW "events" AS
   SELECT
     "bintrail_id", "event_date", "event_hour",
@@ -56,7 +57,7 @@ CREATE OR REPLACE VIEW "events" AS
     "connection_id",
     "schema_name",
     "table_name",
-    CAST("event_type" AS TINYINT) AS "event_type_code",
+    CAST("event_type" AS INTEGER) AS "event_type_code",
     CASE "event_type"
       WHEN 1 THEN 'INSERT'
       WHEN 2 THEN 'UPDATE'
